Lesson 52 Multi-output Categorization ANN
Added and processed the Iris dataset from seaborn for analysis. Utilized a neural network model with PyTorch for multiclass classification, trained it over 1000 epochs using the stochastic gradient descent (SGD) optimizer with the cross-entropy loss function. Recorded loss and accuracy. Visualized the model's training progress, including loss and accuracy over epochs. Achieved a final accuracy of approximately 97% on the Iris dataset, demonstrating the model's ability to classify the species of iris flowers correctly based on the features in the dataset.