Lesson 56 Defining Custom Models
In this lecture, I learned how to create and train a custom classification model with pytorch. The dataset is generated with two clusters, 'a' and 'b,' and the true labels are provided. I've designed a custom neural network model called 'CustomModel' with an input layer, ReLU activation, and an output layer with a sigmoid activation function. The model is trained using binary cross-entropy loss and stochastic gradient descent (SGD) optimizer over 2000 epochs. I've visualized the loss over training and calculated the final accuracy. The accuracy is displayed through a plot, showing correctly classified points in blue and misclassified points in red.